refactor(app): extract SortableJS wiring out of pawwork-sidebar
pawwork-sidebar.tsx grew to 743 lines as the spike's drag patterns
accumulated inline: DOM revert, newDraggableIndex bail, project-group
put rule, no-op short-circuit, plus four blocks of spike-rationale
comments. None of that touches Solid render concerns — it is purely the
SortableJS lifecycle for a div ref.

Move it to pawwork-sidebar-drag.ts as createSortableAttacher(deps):

- deps: onDragSession handler, setIsDragging signal setter,
  getVisiblePinnedIDs accessor.
- Returns the same (kind) => (el) => void shape callers already used,
  so the call site collapses to a single createSortableAttacher
  invocation and three ref={attachSortable('...')} usages.
- onCleanup stays inside the returned ref callback, so the cleanup
  scope (Show branch owner) is unchanged from the inline version.
- The spike's accumulated context now lives in a single file-level
  doc comment instead of being interleaved with JSX.

pawwork-sidebar.tsx: 743 -> 647 lines.
